requirements.txt
requests
beautifulsoup4
click
pandas
python-dotenv
prometheus-client